+++ bindings/nodejs/lib/types/secret_manager/bridge/secret-manager.ts
@@ -42,6 +42,10 @@ export interface __StoreMnemonicMethod__ {
     };
 }
 
+export interface __GetSeedMethod__ {
+    name: 'getSeed';
+}
+
 export interface __SignEd25519Method__ {
     name: 'signEd25519';
     data: {