+++ bindings/nodejs/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.1.5 - 2025-01-07
+
+### Added
+
+- `SecretManager::getSeed()`;
+
 ## 1.1.5 - 2024-01-29
 
 ### Added